% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationOpMatrix.R
\name{CorrelationOpMatrix}
\alias{CorrelationOpMatrix}
\title{Calculates the correlation operator matrix of a covariance matrix \code{M} with respect to the partition \code{partition}.}
\usage{
CorrelationOpMatrix(M, partition, epsilon = 1e-15, type = "norms")
}
\arguments{
\item{M}{A matrix.}

\item{partition}{An integer or a partition (see \code{\link{Partition}}).}

\item{epsilon}{The regularization parameter. By default \code{epsilon == 1e-15}.}

\item{type}{A string equal to \code{"norms"} or \code{"matrices"}. By default \code{type == "norms"}.}
}
\value{
A matrix representing the correlation operator matrix  if \code{type == "matrices"} or a matrix of the operator norms of constituent operators if \code{type == "norms"}.
}
\description{
Calculates the correlation operator matrix of a covariance matrix \code{M} with respect to the partition \code{partition}.
}
\examples{
CorrelationOpMatrix(matrix(1, 10, 10), 3, 1e-10, "norms")
CorrelationOpMatrix(matrix(1, 10, 10), Partition(10, 3), 1e-10, "norms")
}
\seealso{
\code{\link[=PrecisionOpMatrix]{PrecisionOpMatrix()}}
\code{\link[=OpNormMatrix]{OpNormMatrix()}}
}
