% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partition.R
\name{Partition}
\alias{Partition}
\title{Partition \code{1:n} into \code{p} contiguous subvectors of roughly equal length.}
\usage{
Partition(n, p)
}
\arguments{
\item{n}{An integer.}

\item{p}{An integer.}
}
\value{
A list of vectors.
}
\description{
Partition \code{1:n} into \code{p} contiguous subvectors of roughly equal length.
}
\examples{
Partition(10, 3)

}
